+++ src/main/java/com/codeit/weatherwear/domain/feed/entity/FeedComment.java
@@ -1,4 +1,4 @@
-package com.codeit.weatherwear.feed.entity;
+package com.codeit.weatherwear.domain.feed.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -20,7 +20,7 @@ import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @Entity
-@Table(name = "feed_comment")
+@Table(name = "feed_comments")
 @EntityListeners(AuditingEntityListener.class)
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
