+++ src/main/java/com/codeit/weatherwear/domain/feed/entity/FeedLike.java
@@ -1,4 +1,4 @@
-package com.codeit.weatherwear.feed.entity;
+package com.codeit.weatherwear.domain.feed.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -21,7 +21,7 @@ import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 @Entity
-@Table(name = "feed_like")
+@Table(name = "feed_likes")
 @EntityListeners(AuditingEntityListener.class)
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
